requirements.txt
absl-py @ file:///opt/conda/conda-bld/absl-py_1639803114343/work
aiohttp @ file:///tmp/build/80754af9/aiohttp_1646788240124/work
aiosignal @ file:///tmp/build/80754af9/aiosignal_1637843061372/work
async-timeout @ file:///tmp/build/80754af9/async-timeout_1637851218186/work
asynctest==0.13.0
attrs @ file:///opt/conda/conda-bld/attrs_1642510447205/work
blinker==1.4
brotlipy==0.7.0
cachetools @ file:///tmp/build/80754af9/cachetools_1619597386817/work
certifi @ file:///opt/conda/conda-bld/certifi_1655968806487/work/certifi
cffi @ file:///opt/conda/conda-bld/cffi_1642701102775/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click @ file:///tmp/build/80754af9/click_1646056591685/work
cryptography @ file:///tmp/build/80754af9/cryptography_1652083456434/work
cycler==0.11.0
dgl-cu102==0.6.1
fonttools==4.34.4
frozenlist @ file:///tmp/build/80754af9/frozenlist_1637767116045/work
gensim==4.2.0
google-auth @ file:///opt/conda/conda-bld/google-auth_1646735974934/work
google-auth-oauthlib @ file:///tmp/build/80754af9/google-auth-oauthlib_1617120569401/work
grpcio @ file:///tmp/build/80754af9/grpcio_1637578017926/work
idna @ file:///tmp/build/80754af9/idna_1637925883363/work
importlib-metadata @ file:///tmp/build/80754af9/importlib-metadata_1648562407465/work
info-nce-pytorch==0.1.4
isodate==0.6.1
Jinja2==3.1.2
joblib==1.1.0
kiwisolver==1.4.4
littleutils==0.2.2
llvmlite==0.39.0
Markdown @ file:///tmp/build/80754af9/markdown_1614363832606/work
MarkupSafe==2.1.1
matplotlib==3.5.2
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626179032232/work
mkl-service==2.4.0
multidict @ file:///opt/conda/conda-bld/multidict_1640703752579/work
networkx==2.6.3
numba==0.56.0
numpy @ file:///opt/conda/conda-bld/numpy_and_numpy_base_1653915516269/work
oauthlib==3.0.0
ogb==1.3.3
outdated==0.2.1
packaging==21.3
pandas==1.3.5
Pillow==9.2.0
protobuf==3.20.1
psutil==5.9.1
pyasn1 @ file:///Users/ktietz/demo/mc3/conda-bld/pyasn1_1629708007385/work
pyasn1-modules==0.2.8
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
PyJWT @ file:///opt/conda/conda-bld/pyjwt_1657544592787/work
pyOpenSSL @ file:///opt/conda/conda-bld/pyopenssl_1643788558760/work
pyparsing==3.0.9
PySocks @ file:///tmp/build/80754af9/pysocks_1594394576006/work
python-dateutil==2.8.2
pytz==2022.1
rdflib==6.2.0
requests @ file:///opt/conda/conda-bld/requests_1657734628632/work
requests-oauthlib==1.3.0
rsa @ file:///tmp/build/80754af9/rsa_1614366226499/work
scikit-learn==1.0.2
scipy==1.7.3
six @ file:///tmp/build/80754af9/six_1644875935023/work
smart-open==6.0.0
tensorboard @ file:///home/builder/tkoch/workspace/tensorflow/tensorboard_1658918251247/work/tensorboard-2.8.0-py3-none-any.whl
tensorboard-data-server @ file:///tmp/build/80754af9/tensorboard-data-server_1633024562496/work/tensorboard_data_server-0.6.0-py3-none-manylinux2010_x86_64.whl
tensorboard-plugin-wit @ file:///home/builder/tkoch/workspace/tensorflow/tensorboard-plugin-wit_1658918494740/work/tensorboard_plugin_wit-1.8.1-py3-none-any.whl
threadpoolctl==3.1.0
torch==1.12.0
torch-cluster==1.6.0
torch-geometric==2.0.4
torch-scatter==2.0.9
torch-sparse==0.6.14
torchaudio==0.12.0
torchvision==0.13.0
tqdm==4.64.0
typing_extensions @ file:///opt/conda/conda-bld/typing_extensions_1647553014482/work
urllib3 @ file:///opt/conda/conda-bld/urllib3_1650639997961/work
Werkzeug @ file:///opt/conda/conda-bld/werkzeug_1645628268370/work
word2vec==0.11.1
yarl @ file:///tmp/build/80754af9/yarl_1606939952652/work
zipp @ file:///opt/conda/conda-bld/zipp_1652341764480/work
